scanner is moved out of its crappy separate folder
