+++ tests/ShuntingYard/Makefile
@@ -4,8 +4,8 @@ a.exe: main.o
 main.o: tests/ShuntingYard/main.cpp PostFixEvaluator.o
 	g++ -c -std=c++20 tests/ShuntingYard/main.cpp -o main.o
 
-Scanner.o: Scanner/Scanner.hpp Scanner/Scanner.cpp SourceReader.o
-	g++ -c -std=c++20 Scanner/Scanner.cpp -o Scanner.o
+Scanner.o: Scanner.hpp Scanner.cpp SourceReader.o
+	g++ -c -std=c++20 Scanner.cpp -o Scanner.o
 
 PostFixEvaluator.o: Scanner.o PostFixEvaluator.hpp PostFixEvaluator.cpp
 	g++ -c -std=c++20 PostFixEvaluator.cpp -o PostFixEvaluator.o
